feat: implemented activate and deactivate user

+++ src/app/(adminUi)/users/usersTable.tsx
@@ -7,13 +7,13 @@ import Link from "next/link";
 import { Delete, Edit, Eye } from "@/assets";
 import { Modal, Select } from "flowbite-react";
 import Table from "@/components/Table";
-import { IUsersTableProps } from "@/models/User";
+import { IUsersTableProps,  IUser } from "@/models/User";
 import { RiErrorWarningLine } from "react-icons/ri";
 import Button from "@/components/Button";
-import { useDeleteUserMutation } from "./UserApiSlice";
+import { useDeleteUserMutation, useActivateDeactivateUserMutation } from "./UserApiSlice";
 import SuccessModal from "@/components/Modals/SuccessModal";
 import { toast } from "react-toastify";
-
+import { GpToast } from '@/components/Toast';
 
 const PromptModal = (
   bodyText: string,
@@ -50,6 +50,43 @@ const PromptModal = (
   );
 };
 
+export const ConfirmationModal = (
+  bodyText: string,
+  cta: string,
+  showModal: boolean,
+  onClose: () => void,
+  onDone: () => void,
+  doneBtnBusy: boolean
+) => {
+  return (
+    <Modal size="sm" dismissible popup show={showModal} onClose={() => onClose}>
+      <Modal.Body className="h-[300px] p-4 flex  items-center justify-center">
+        <div className=" flex flex-col justify-around items-center text-center">
+          <p>
+            <RiErrorWarningLine color="red" size={34} />
+          </p>
+          <p>{bodyText}</p>
+          <div className=" w-full flex justify-between mt-4 mb-4">
+            <Button appButtonType="grey-button" className="w-[45%]" onClick={onClose}>
+              Cancel
+            </Button>
+
+            <Button
+              isLoading={doneBtnBusy}
+              disabled={doneBtnBusy}
+              appButtonType="red-button"
+              className="w-[45%]"
+              onClick={onDone}
+            >
+              Yes
+            </Button>
+          </div>
+        </div>
+      </Modal.Body>
+    </Modal>
+  );
+};
+
 const UsersTable: React.FC<IUsersTableProps> = ({
   data,
   refetch,
@@ -60,10 +97,22 @@ const UsersTable: React.FC<IUsersTableProps> = ({
   const [deleteUser, deleteResponse] = useDeleteUserMutation();
 
   const [showIsDeleteModal, setShowIsDeleteModal] = useState(false);
+  const [deactivateActivateUser, deactivateActiveResponse] = useActivateDeactivateUserMutation();
   const handleDeleteModal = () => {
     setShowIsDeleteModal(!showIsDeleteModal);
   };
   const [toggleSuccessModal, setToggleSuccessModal] = useState(false);
+  const [showIsDeactivateModal, setShowIsDeactivateModal] = useState<boolean>(false);
+  const [showIsActivateModal, setShowIsActivateModal] = useState<boolean>(false);
+
+
+  const handleDeactivateModal = () => {
+    setShowIsDeactivateModal(!showIsDeactivateModal);
+  };
+  const handleActivateModal = () => {
+    setShowIsActivateModal(!showIsActivateModal);
+  };
+
   const handleDelete = (row: any) => {
     deleteUser(row.id)
       .then((res: any) => {
@@ -79,6 +128,37 @@ const UsersTable: React.FC<IUsersTableProps> = ({
         });
       });
   };
+
+  const handleDeactivateUser = (row: IUser) => {
+    deactivateActivateUser({
+      id: row.id,
+      userId: row.id,
+      isActive: !row.isActive,
+    }).then((res: any) => {
+      if (res.error) {
+        handleDeactivateModal();
+        GpToast({
+          type: 'error',
+          message: 'An error occured, kindly try again',
+          placement: toast.POSITION.TOP_LEFT,
+        });
+      }
+    });
+  };
+
+  useEffect(() => {
+    if (deactivateActiveResponse.isSuccess) {
+      handleDeactivateModal();
+      const { message } = deactivateActiveResponse.data;
+      GpToast({
+        type: 'success',
+        message: message,
+        placement: toast.POSITION.TOP_LEFT,
+      });
+      refetch();
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [deactivateActiveResponse]);
 
   useEffect(() => {
     if (deleteResponse.isSuccess) {
@@ -113,7 +193,7 @@ const UsersTable: React.FC<IUsersTableProps> = ({
 
   const handleClick = (index: number) => {
     const newVisibilities = [...visibilities];
-    newVisibilities.map((thisVisibility, ind) => {
+    newVisibilities.map((_thisVisibility, ind) => {
       index !== ind
         ? (newVisibilities[ind] = false)
         : (newVisibilities[index] = !newVisibilities[index]);
@@ -153,7 +233,7 @@ const UsersTable: React.FC<IUsersTableProps> = ({
                 {row.isActive ? "Active" : "Inactive"}
               </div>
             ),
-            option: (
+           option: (
               <div className="">
                 <div className="relative">
                   <div className="flex items-center cursor-pointer">
@@ -162,11 +242,11 @@ const UsersTable: React.FC<IUsersTableProps> = ({
                   {visibilities && visibilities[index] ? (
                     <div className="absolute border border-muted rounded-md z-10 right-0 top-full px-3 w-max bg-white">
                       <Link
-                        href={`/users/${row.id}`}
+                        href={`/application-management/users/${row.id}`}
                         className="flex cursor-pointer text-left py-3 border-b border-neutral-50 text-small text-gray-700 items-center gap-2"
                       >
                         <Eye />
-                        View Profile
+                        View
                       </Link>
                       <div
                         onClick={() => {
@@ -175,7 +255,7 @@ const UsersTable: React.FC<IUsersTableProps> = ({
                         className="flex cursor-pointer text-left py-3 border-b border-neutral-50 text-small text-gray-700 items-center gap-2"
                       >
                         <Edit />
-                        Edit Profile
+                        Edit
                       </div>
                       <div
                         onClick={() => handleDeleteModal()}
@@ -183,32 +263,64 @@ const UsersTable: React.FC<IUsersTableProps> = ({
                       >
                         <Delete />
                         Delete
+                        {ConfirmationModal(
+                          'Are you sure you want to delete this user?',
+                          'Delete',
+                          showIsDeleteModal,
+
+                          () => {
+                            setShowIsDeleteModal(!showIsDeleteModal);
+                          },
+                          () => {
+                            handleDelete(row);
+                          },
+                          deleteResponse.isLoading
+                        )}
                       </div>
-                      {PromptModal(
-                        "Are sure you want to delete this user?",
-                        showIsDeleteModal,
-                        () => {
-                          setShowIsDeleteModal(!showIsDeleteModal);
-                        },
-                        () => {
-                          handleDelete(row);
-                        },
-                        deleteResponse.isLoading
-                      )}
-                      {toggleSuccessModal && (
-                        <SuccessModal
-                          onDoneClicked={() =>
-                            setToggleSuccessModal(!toggleSuccessModal)
-                          }
-                          message="User is deleted
-                     Successfully"
-                        />
+                      {row.isActive ? (
+                        <div
+                          onClick={() => handleDeactivateModal()}
+                          className="flex cursor-pointer text-left py-3 border-b border-neutral-50 text-small text-red-600 items-center gap-2"
+                        >
+                          Deactivate
+                          {ConfirmationModal(
+                            'Are you sure you want to deactivate this user?',
+                            'Deactivate',
+                            showIsDeactivateModal,
+                            () => {
+                              setShowIsDeactivateModal(!showIsDeactivateModal);
+                            },
+                            () => {
+                              handleDeactivateUser(row);
+                            },
+                            deactivateActiveResponse.isLoading
+                          )}
+                        </div>
+                      ) : (
+                        <div
+                          onClick={() => handleActivateModal()}
+                          className="flex cursor-pointer text-left py-3 border-b border-neutral-50 text-small text-gray-700 items-center gap-2"
+                        >
+                          Activate
+                          {ConfirmationModal(
+                            'Are you sure you want to activate this user?',
+                            'Activate',
+                            showIsActivateModal,
+                            () => {
+                              setShowIsActivateModal(!showIsActivateModal);
+                            },
+                            () => {
+                              handleDeactivateUser(row);
+                            },
+                            deactivateActiveResponse.isLoading
+                          )}
+                        </div>
                       )}
                     </div>
                   ) : (
-                    ""
-                  )}{" "}
-                </div>{" "}
+                    ''
+                  )}
+                </div>
               </div>
             ),
           };
